+++ finetune/target_models/sequence_labeling.py
@@ -78,7 +78,7 @@ class SequencePipeline(BasePipeline):
             return (
                 (
                     {"tokens": tf.int32, "mask": tf.float32, "context": tf.float32},
-                    tf.int32,
+                    tf.float32,
                 ),
                 (
                     {
@@ -91,7 +91,7 @@ class SequencePipeline(BasePipeline):
             )
         else:
             return (
-                ({"tokens": tf.int32, "mask": tf.float32}, tf.int32),
+                ({"tokens": tf.int32, "mask": tf.float32}, tf.float32),
                 (
                     {
                         "tokens": TS([self.config.max_length, 2]),
